First try at converting user-manual to AsciiDoc
You can do "make doc" in the main directory to create the html version,
and if you want to play around with it, do "make show" in the
Documentation subdirectory to start firefox on the end result.

It's by no means perfect, but it gives somewhat reasonable results, and
this is enough initial work for people to play around with, I think.

NOTE! You need "asciidoc" installed to do this: it's a python program,
so it should be pretty easy even on non-Linux platforms.  And on Linux,
most distributions package it, so you just have to do something like

	yum install asciidoc

to get it (replace with apt-get/zypper/whatever).

Asciidoc can generate other output too (man-pages, LaTeX, etc), maybe
people want to play with that part too.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>