Added a preference for time format
Added a preference so that the users can changed the time format to
a strftime format string of their own choosing. The preference has its
own frame named "Time format (http://linux.die.net/man/3/strftime)"
between the frame "Columns" and the font button.

I implemented usage of the global time format in the places where I
found a date that was outputted to the user. These places were the
divelist, the statistics, the dive info and the output when printing.

Signed-off-by: Johannes Larsen <johs.a.larsen@gmail.com>