Cochran: fix up dive data descrambling
This seems to do the dive data descrambling right for both files I have
access to.  Except it uses a hardcoded (different) offset for the two.
I have yet to figure out how to automatically detect the offset itself
properly, so you have to compile for the right file.

I'll figure it out, but I'm committing this as a reasonable point in the
process.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>