divecomputer importing: show the date of the currently importing dive
I'm hoping most other dive computers are quicker to import from than the
Suunto I have, but mine can take minutes to import all the dives.  Sure,
we have that nice progress bar, so it shows that it's doing something,
but it's not really showing *what* it is doing.

So instead of showing just "Parsing dive X", let's show the date of the
dive.  That way, when it takes two minutes to import all the dives, at
least you can see "oh, it's going back to the dives of last year" and it
then feels like you have some good reason for the delay.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>