Fix dive notes import from Suundo SDM
The xslt translation didn't add the <notes> tag for the notes, so while
it did select the notes from the SDM file, that never made it into the
subsurface notes.

Also added weather info to the notes, mainly as an example.

There are probably other things we could do, but this fixes at least the
trivial test-case from Szymon Kosecki.

Reported-by: Szymon Kosecki <skosecki@gmail.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>