Suunto SDE conversion: add boat name to notes if it exists
This is, I think, the last piece of relevant information that I can find
in Szymon's SDE file.

Which is not to mean that we get all the conversions right, or that we
handle the more complex cases (still no multi-cylinder import, for
example). But it should be much better than it used to be.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>