Allow overriding the default xslt path
It's very annoying to have to do "make install" to test a new xslt file,
just because the default xslt path has the standard install path as the
first entry.

At the same time, we do want to default to just using the standard
install location first.

So to allow both testing, and having a nice sane default, just add
support for a SUBSURFACE_XSLT_PATH environment variable that overrides
the default one if it exists.

So then you can just do

   SUBSURFACE_XSLT_PATH=xslt ./subsurface

to run subsurface from inside the git tree itself, using the current
files in the git xslt subdirectory.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>